Added tail bones to catGun.blender
Added tail armature bones to the catGun blender file that can be animated later. They follow the same transformations as the rest of the body since they're part of the same armature. TO do this, I added a bone to the base of the skeleton, made the tail, then deleted the connector bone near the base.