Implemented a path with yuka AI
I implemented a path with yuka, but the test circle seems to leave the area it's suppsoed to after a few cycles; the circle just starts to drift off while still doing them. I have also discovered that making the next waypoint for the FollowPathBehavior constructor makes the 'vehicle' very jittery if it starts to go above the default value 1.