Cat moves while airborne
The cat object had trouble moving on the ground with its bounding box, but without the friction of the ground it easily moves on its path set with yuka. There's also a temporary variable in the moveacat function that helps with the glifht of the cacat, and will be removed later.