Cat moves cleanly on path
In the moveACat function, the resultantImpulse is now set so that the force of the box moves to match the position of the yuka vehicle on the path. One downside is that the cat or the box can't be pushed off anymore. We can jsut say the cat's so buff that it refuses to move. The rotation of the cat isn't matched yet.