Added pursuit behavior
The cat now chases the player slowly, and tries to push the player off of things. I may need to implement an offsetPursuit behavior so that it stays a certain distance away, then I need to figure out a way to detect that it's in shooting range. There is also a movingEntity object following the player closely. I also removed some unused variables and comments in a few functions.